+++ game_patch/multi/server.h
@@ -12,3 +12,4 @@ bool check_server_chat_command(const char* msg, rf::Player* sender);
 bool server_is_saving_enabled();
 void server_reliable_socket_ready(rf::Player* player);
 bool server_weapon_items_give_full_ammo();
+void process_queued_spawn_points_from_items();