+++ game_patch/rf/math/vector.h
@@ -16,6 +16,15 @@ namespace rf
 
         bool operator==(const Vector3& other) const = default;
 
+        [[nodiscard]] Vector3 cross(const Vector3& other) const
+        {
+            return Vector3{
+                y * other.z - z * other.y,
+                z * other.x - x * other.z,
+                x * other.y - y * other.x
+            };
+        }
+
         Vector3& operator+=(const Vector3& other)
         {
             x += other.x;
@@ -171,4 +180,5 @@ namespace rf
     };
 
     static auto& vec2_zero_vector = addr_as_ref<Vector2>(0x0173C370);
+    static auto& vec_dist_squared = addr_as_ref<float(const rf::Vector3*, const rf::Vector3*)>(0x004FAF00);
 }