+++ docs/CHANGELOG.md
@@ -44,6 +44,9 @@ Version 1.9.0 (not released yet)
 - Add level filename to "Level Initializing" console message
 - Properly handle WM_PAINT in dedicated server, may improve performance (DF bug)
 - Fix crash when `verify_level` command is run without a level being loaded
+- Make game spawn players at Player Start position if level has no respawn points
+- Add respawn point selection logic settings to dedicated server config
+- Raise limit on number of multiplayer respawn points per level from 32 to 2048
 
 Version 1.8.0 (released 2022-09-17)
 -----------------------------------